Move Characteristic and Constraint interfaces to proper packages
