+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/characteristic/impl/DefaultCode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Robert Bosch Manufacturing Solutions GmbH
+ * Copyright (c) 2022 Robert Bosch Manufacturing Solutions GmbH
  *
  * See the AUTHORS file(s) distributed with this work for additional
  * information regarding authorship.
@@ -11,28 +11,21 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-package io.openmanufacturing.sds.metamodel.impl;
+package io.openmanufacturing.sds.characteristic.impl;
 
 import java.util.Optional;
 import java.util.StringJoiner;
 
-import io.openmanufacturing.sds.metamodel.Characteristic;
-import io.openmanufacturing.sds.characteristic.Set;
-import io.openmanufacturing.sds.metamodel.CollectionValue;
+import io.openmanufacturing.sds.characteristic.Code;
 import io.openmanufacturing.sds.metamodel.Type;
+import io.openmanufacturing.sds.metamodel.impl.DefaultCharacteristic;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.visitor.AspectVisitor;
 
-public class DefaultSet extends DefaultCollection implements Set {
+public class DefaultCode extends DefaultCharacteristic implements Code {
 
-   public DefaultSet( final MetaModelBaseAttributes metaModelBaseAttributes,
-         final Optional<Type> dataType, final Optional<Characteristic> elementCharacteristic ) {
-      super( metaModelBaseAttributes, dataType, false, false, elementCharacteristic );
-   }
-
-   @Override
-   public CollectionValue.CollectionType getCollectionType() {
-      return CollectionValue.CollectionType.SET;
+   public DefaultCode( final MetaModelBaseAttributes metaModelBaseAttributes, final Type dataType ) {
+      super( metaModelBaseAttributes, Optional.of( dataType ) );
    }
 
    /**
@@ -44,12 +37,12 @@ public class DefaultSet extends DefaultCollection implements Set {
     */
    @Override
    public <T, C> T accept( final AspectVisitor<T, C> visitor, final C context ) {
-      return visitor.visitSet( this, context );
+      return visitor.visitCode( this, context );
    }
 
    @Override
    public String toString() {
-      return new StringJoiner( ", ", DefaultSet.class.getSimpleName() + "[", "]" )
+      return new StringJoiner( ", ", DefaultCode.class.getSimpleName() + "[", "]" )
             .toString();
    }
 }
