+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/characteristic/impl/DefaultEnumeration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Robert Bosch Manufacturing Solutions GmbH
+ * Copyright (c) 2022 Robert Bosch Manufacturing Solutions GmbH
  *
  * See the AUTHORS file(s) distributed with this work for additional
  * information regarding authorship.
@@ -10,34 +10,37 @@
  *
  * SPDX-License-Identifier: MPL-2.0
  */
-package io.openmanufacturing.sds.metamodel.impl;
+package io.openmanufacturing.sds.characteristic.impl;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 import java.util.StringJoiner;
 
-import io.openmanufacturing.sds.characteristic.State;
+import io.openmanufacturing.sds.characteristic.Enumeration;
 import io.openmanufacturing.sds.metamodel.Type;
 import io.openmanufacturing.sds.metamodel.Value;
+import io.openmanufacturing.sds.metamodel.impl.DefaultCharacteristic;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.visitor.AspectVisitor;
 
-public class DefaultState extends DefaultEnumeration implements State {
-   private final Value defaultValue;
+@SuppressWarnings( "squid:S1150" ) // Sonar thinks this implements java.util.Enumeration, which it does not
+public class DefaultEnumeration extends DefaultCharacteristic implements Enumeration {
+   private final List<Value> values;
 
-   public DefaultState( final MetaModelBaseAttributes metaModelBaseAttributes, final Type dataType, final List<Value> values, final Value defaultValue ) {
-      super( metaModelBaseAttributes, dataType, values );
-      this.defaultValue = defaultValue;
+   public DefaultEnumeration( final MetaModelBaseAttributes metaModelBaseAttributes, final Type dataType, final List<Value> values ) {
+      super( metaModelBaseAttributes, Optional.of( dataType ) );
+      this.values = values;
    }
 
    /**
-    * The default value for this state
+    * A list of valid states.
     *
-    * @return the defaultValue.
+    * @return the values.
     */
    @Override
-   public Value getDefaultValue() {
-      return defaultValue;
+   public List<Value> getValues() {
+      return values;
    }
 
    /**
@@ -49,13 +52,13 @@ public class DefaultState extends DefaultEnumeration implements State {
     */
    @Override
    public <T, C> T accept( final AspectVisitor<T, C> visitor, final C context ) {
-      return visitor.visitState( this, context );
+      return visitor.visitEnumeration( this, context );
    }
 
    @Override
    public String toString() {
-      return new StringJoiner( ", ", DefaultState.class.getSimpleName() + "[", "]" )
-            .add( "defaultValue=" + defaultValue )
+      return new StringJoiner( ", ", DefaultEnumeration.class.getSimpleName() + "[", "]" )
+            .add( "values=" + values )
             .toString();
    }
 
@@ -70,12 +73,12 @@ public class DefaultState extends DefaultEnumeration implements State {
       if ( !super.equals( o ) ) {
          return false;
       }
-      final DefaultState that = (DefaultState) o;
-      return Objects.equals( defaultValue, that.defaultValue );
+      final DefaultEnumeration that = (DefaultEnumeration) o;
+      return Objects.equals( values, that.values );
    }
 
    @Override
    public int hashCode() {
-      return Objects.hash( super.hashCode(), defaultValue );
+      return Objects.hash( super.hashCode(), values );
    }
 }
