+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/characteristic/impl/DefaultMeasurement.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Robert Bosch Manufacturing Solutions GmbH
+ * Copyright (c) 2022 Robert Bosch Manufacturing Solutions GmbH
  *
  * See the AUTHORS file(s) distributed with this work for additional
  * information regarding authorship.
@@ -10,20 +10,20 @@
  *
  * SPDX-License-Identifier: MPL-2.0
  */
-package io.openmanufacturing.sds.metamodel.impl;
+package io.openmanufacturing.sds.characteristic.impl;
 
 import java.util.Optional;
 import java.util.StringJoiner;
 
-import io.openmanufacturing.sds.characteristic.Duration;
+import io.openmanufacturing.sds.characteristic.Measurement;
 import io.openmanufacturing.sds.metamodel.Type;
 import io.openmanufacturing.sds.metamodel.Unit;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.visitor.AspectVisitor;
 
-public class DefaultDuration extends DefaultQuantifiable implements Duration {
+public class DefaultMeasurement extends DefaultQuantifiable implements Measurement {
 
-   public DefaultDuration( final MetaModelBaseAttributes metaModelBaseAttributes, final Type dataType, final Optional<Unit> unit ) {
+   public DefaultMeasurement( final MetaModelBaseAttributes metaModelBaseAttributes, final Type dataType, final Optional<Unit> unit ) {
       super( metaModelBaseAttributes, dataType, unit );
    }
 
@@ -36,12 +36,12 @@ public class DefaultDuration extends DefaultQuantifiable implements Duration {
     */
    @Override
    public <T, C> T accept( final AspectVisitor<T, C> visitor, final C context ) {
-      return visitor.visitDuration( this, context );
+      return visitor.visitMeasurement( this, context );
    }
 
    @Override
    public String toString() {
-      return new StringJoiner( ", ", DefaultDuration.class.getSimpleName() + "[", "]" )
+      return new StringJoiner( ", ", DefaultMeasurement.class.getSimpleName() + "[", "]" )
             .toString();
    }
 }
