+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/characteristic/impl/DefaultQuantifiable.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Robert Bosch Manufacturing Solutions GmbH
+ * Copyright (c) 2022 Robert Bosch Manufacturing Solutions GmbH
  *
  * See the AUTHORS file(s) distributed with this work for additional
  * information regarding authorship.
@@ -10,36 +10,35 @@
  *
  * SPDX-License-Identifier: MPL-2.0
  */
-package io.openmanufacturing.sds.metamodel.impl;
+package io.openmanufacturing.sds.characteristic.impl;
 
-import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.StringJoiner;
 
-import io.openmanufacturing.sds.characteristic.Enumeration;
+import io.openmanufacturing.sds.characteristic.Quantifiable;
 import io.openmanufacturing.sds.metamodel.Type;
-import io.openmanufacturing.sds.metamodel.Value;
+import io.openmanufacturing.sds.metamodel.Unit;
+import io.openmanufacturing.sds.metamodel.impl.DefaultCharacteristic;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.visitor.AspectVisitor;
 
-@SuppressWarnings( "squid:S1150" ) // Sonar thinks this implements java.util.Enumeration, which it does not
-public class DefaultEnumeration extends DefaultCharacteristic implements Enumeration {
-   private final List<Value> values;
+public class DefaultQuantifiable extends DefaultCharacteristic implements Quantifiable {
+   private final Optional<Unit> unit;
 
-   public DefaultEnumeration( final MetaModelBaseAttributes metaModelBaseAttributes, final Type dataType, final List<Value> values ) {
+   public DefaultQuantifiable( final MetaModelBaseAttributes metaModelBaseAttributes, final Type dataType, final Optional<Unit> unit ) {
       super( metaModelBaseAttributes, Optional.of( dataType ) );
-      this.values = values;
+      this.unit = unit;
    }
 
    /**
-    * A list of valid states.
+    * The Unit of the Quantifiable.
     *
-    * @return the values.
+    * @return the unit.
     */
    @Override
-   public List<Value> getValues() {
-      return values;
+   public Optional<Unit> getUnit() {
+      return unit;
    }
 
    /**
@@ -51,13 +50,13 @@ public class DefaultEnumeration extends DefaultCharacteristic implements Enumera
     */
    @Override
    public <T, C> T accept( final AspectVisitor<T, C> visitor, final C context ) {
-      return visitor.visitEnumeration( this, context );
+      return visitor.visitQuantifiable( this, context );
    }
 
    @Override
    public String toString() {
-      return new StringJoiner( ", ", DefaultEnumeration.class.getSimpleName() + "[", "]" )
-            .add( "values=" + values )
+      return new StringJoiner( ", ", DefaultQuantifiable.class.getSimpleName() + "[", "]" )
+            .add( "unit=" + unit )
             .toString();
    }
 
@@ -72,12 +71,12 @@ public class DefaultEnumeration extends DefaultCharacteristic implements Enumera
       if ( !super.equals( o ) ) {
          return false;
       }
-      final DefaultEnumeration that = (DefaultEnumeration) o;
-      return Objects.equals( values, that.values );
+      final DefaultQuantifiable that = (DefaultQuantifiable) o;
+      return Objects.equals( unit, that.unit );
    }
 
    @Override
    public int hashCode() {
-      return Objects.hash( super.hashCode(), values );
+      return Objects.hash( super.hashCode(), unit );
    }
 }
