+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/characteristic/impl/DefaultSet.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 Robert Bosch Manufacturing Solutions GmbH
+ * Copyright (c) 2022 Robert Bosch Manufacturing Solutions GmbH
  *
  * See the AUTHORS file(s) distributed with this work for additional
  * information regarding authorship.
@@ -10,21 +10,29 @@
  *
  * SPDX-License-Identifier: MPL-2.0
  */
-package io.openmanufacturing.sds.metamodel.impl;
+
+package io.openmanufacturing.sds.characteristic.impl;
 
 import java.util.Optional;
 import java.util.StringJoiner;
 
-import io.openmanufacturing.sds.characteristic.Measurement;
+import io.openmanufacturing.sds.metamodel.Characteristic;
+import io.openmanufacturing.sds.characteristic.Set;
+import io.openmanufacturing.sds.metamodel.CollectionValue;
 import io.openmanufacturing.sds.metamodel.Type;
-import io.openmanufacturing.sds.metamodel.Unit;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.visitor.AspectVisitor;
 
-public class DefaultMeasurement extends DefaultQuantifiable implements Measurement {
+public class DefaultSet extends DefaultCollection implements Set {
 
-   public DefaultMeasurement( final MetaModelBaseAttributes metaModelBaseAttributes, final Type dataType, final Optional<Unit> unit ) {
-      super( metaModelBaseAttributes, dataType, unit );
+   public DefaultSet( final MetaModelBaseAttributes metaModelBaseAttributes,
+         final Optional<Type> dataType, final Optional<Characteristic> elementCharacteristic ) {
+      super( metaModelBaseAttributes, dataType, false, false, elementCharacteristic );
+   }
+
+   @Override
+   public CollectionValue.CollectionType getCollectionType() {
+      return CollectionValue.CollectionType.SET;
    }
 
    /**
@@ -36,12 +44,12 @@ public class DefaultMeasurement extends DefaultQuantifiable implements Measureme
     */
    @Override
    public <T, C> T accept( final AspectVisitor<T, C> visitor, final C context ) {
-      return visitor.visitMeasurement( this, context );
+      return visitor.visitSet( this, context );
    }
 
    @Override
    public String toString() {
-      return new StringJoiner( ", ", DefaultMeasurement.class.getSimpleName() + "[", "]" )
+      return new StringJoiner( ", ", DefaultSet.class.getSimpleName() + "[", "]" )
             .toString();
    }
 }
