+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/characteristic/impl/DefaultSingleEntity.java
@@ -1,8 +1,8 @@
 /*
- * Copyright (c) 2021 Robert Bosch Manufacturing Solutions GmbH
+ * Copyright (c) 2022 Robert Bosch Manufacturing Solutions GmbH
  *
  * See the AUTHORS file(s) distributed with this work for additional
- * information regarding authorship. 
+ * information regarding authorship.
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -11,28 +11,21 @@
  * SPDX-License-Identifier: MPL-2.0
  */
 
-package io.openmanufacturing.sds.metamodel.impl;
+package io.openmanufacturing.sds.characteristic.impl;
 
 import java.util.Optional;
 import java.util.StringJoiner;
 
-import io.openmanufacturing.sds.metamodel.Characteristic;
-import io.openmanufacturing.sds.characteristic.List;
-import io.openmanufacturing.sds.metamodel.CollectionValue;
+import io.openmanufacturing.sds.characteristic.SingleEntity;
 import io.openmanufacturing.sds.metamodel.Type;
+import io.openmanufacturing.sds.metamodel.impl.DefaultCharacteristic;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.visitor.AspectVisitor;
 
-public class DefaultList extends DefaultCollection implements List {
+public class DefaultSingleEntity extends DefaultCharacteristic implements SingleEntity {
 
-   public DefaultList( final MetaModelBaseAttributes metaModelBaseAttributes,
-         final Optional<Type> dataType, final Optional<Characteristic> elementCharacteristic ) {
-      super( metaModelBaseAttributes, dataType, true, true, elementCharacteristic );
-   }
-
-   @Override
-   public CollectionValue.CollectionType getCollectionType() {
-      return CollectionValue.CollectionType.LIST;
+   public DefaultSingleEntity( final MetaModelBaseAttributes metaModelBaseAttributes, final Type dataType ) {
+      super( metaModelBaseAttributes, Optional.of( dataType ) );
    }
 
    /**
@@ -44,12 +37,12 @@ public class DefaultList extends DefaultCollection implements List {
     */
    @Override
    public <T, C> T accept( final AspectVisitor<T, C> visitor, final C context ) {
-      return visitor.visitList( this, context );
+      return visitor.visitSingleEntity( this, context );
    }
 
    @Override
    public String toString() {
-      return new StringJoiner( ", ", DefaultList.class.getSimpleName() + "[", "]" )
+      return new StringJoiner( ", ", DefaultSingleEntity.class.getSimpleName() + "[", "]" )
             .toString();
    }
 }
