+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/constraint/impl/DefaultEncodingConstraint.java
@@ -1,8 +1,8 @@
 /*
- * Copyright (c) 2021 Robert Bosch Manufacturing Solutions GmbH
+ * Copyright (c) 2022 Robert Bosch Manufacturing Solutions GmbH
  *
  * See the AUTHORS file(s) distributed with this work for additional
- * information regarding authorship. 
+ * information regarding authorship.
  *
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -10,32 +10,33 @@
  *
  * SPDX-License-Identifier: MPL-2.0
  */
-package io.openmanufacturing.sds.metamodel.impl;
+package io.openmanufacturing.sds.constraint.impl;
 
-import java.util.Locale;
+import java.nio.charset.Charset;
 import java.util.Objects;
 import java.util.StringJoiner;
 
-import io.openmanufacturing.sds.constraint.LanguageConstraint;
+import io.openmanufacturing.sds.constraint.EncodingConstraint;
+import io.openmanufacturing.sds.metamodel.impl.DefaultConstraint;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.visitor.AspectVisitor;
 
-public class DefaultLanguageConstraint extends DefaultConstraint implements LanguageConstraint {
-   private final Locale languageCode;
+public class DefaultEncodingConstraint extends DefaultConstraint implements EncodingConstraint {
+   private final Charset value;
 
-   public DefaultLanguageConstraint( final MetaModelBaseAttributes metaModelBaseAttributes, final Locale languageCode ) {
+   public DefaultEncodingConstraint( final MetaModelBaseAttributes metaModelBaseAttributes, final Charset value ) {
       super( metaModelBaseAttributes );
-      this.languageCode = languageCode;
+      this.value = value;
    }
 
    /**
-    * An ISO 639-1 language code for the language of the value of the constrained Property
+    * Constrains the encoding (character set) of a property.
     *
-    * @return the languageCode.
+    * @return the value.
     */
    @Override
-   public Locale getLanguageCode() {
-      return languageCode;
+   public Charset getValue() {
+      return value;
    }
 
    /**
@@ -47,13 +48,13 @@ public class DefaultLanguageConstraint extends DefaultConstraint implements Lang
     */
    @Override
    public <T, C> T accept( final AspectVisitor<T, C> visitor, final C context ) {
-      return visitor.visitLanguageConstraint( this, context );
+      return visitor.visitEncodingConstraint( this, context );
    }
 
    @Override
    public String toString() {
-      return new StringJoiner( ", ", DefaultLanguageConstraint.class.getSimpleName() + "[", "]" )
-            .add( "languageCode=" + languageCode )
+      return new StringJoiner( ", ", DefaultEncodingConstraint.class.getSimpleName() + "[", "]" )
+            .add( "value=" + value )
             .toString();
    }
 
@@ -68,12 +69,12 @@ public class DefaultLanguageConstraint extends DefaultConstraint implements Lang
       if ( !super.equals( o ) ) {
          return false;
       }
-      final DefaultLanguageConstraint that = (DefaultLanguageConstraint) o;
-      return Objects.equals( languageCode, that.languageCode );
+      final DefaultEncodingConstraint that = (DefaultEncodingConstraint) o;
+      return Objects.equals( value, that.value );
    }
 
    @Override
    public int hashCode() {
-      return Objects.hash( super.hashCode(), languageCode );
+      return Objects.hash( super.hashCode(), value );
    }
 }
