+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/metamodel/loader/instantiator/DurationInstantiator.java
@@ -21,7 +21,7 @@ import org.apache.jena.rdf.model.Statement;
 import io.openmanufacturing.sds.characteristic.Duration;
 import io.openmanufacturing.sds.metamodel.Type;
 import io.openmanufacturing.sds.metamodel.Unit;
-import io.openmanufacturing.sds.metamodel.impl.DefaultDuration;
+import io.openmanufacturing.sds.characteristic.impl.DefaultDuration;
 import io.openmanufacturing.sds.metamodel.loader.Instantiator;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.loader.ModelElementFactory;
