+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/metamodel/loader/instantiator/LanguageConstraintInstantiator.java
@@ -18,7 +18,7 @@ import java.util.Locale;
 import org.apache.jena.rdf.model.Resource;
 
 import io.openmanufacturing.sds.constraint.LanguageConstraint;
-import io.openmanufacturing.sds.metamodel.impl.DefaultLanguageConstraint;
+import io.openmanufacturing.sds.constraint.impl.DefaultLanguageConstraint;
 import io.openmanufacturing.sds.metamodel.loader.Instantiator;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.loader.ModelElementFactory;
