+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/metamodel/loader/instantiator/LengthConstraintInstantiator.java
@@ -21,7 +21,7 @@ import org.apache.jena.rdf.model.Resource;
 import org.apache.jena.rdf.model.Statement;
 
 import io.openmanufacturing.sds.constraint.LengthConstraint;
-import io.openmanufacturing.sds.metamodel.impl.DefaultLengthConstraint;
+import io.openmanufacturing.sds.constraint.impl.DefaultLengthConstraint;
 import io.openmanufacturing.sds.metamodel.loader.Instantiator;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.loader.ModelElementFactory;
