+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/metamodel/loader/instantiator/LocaleConstraintInstantiator.java
@@ -18,7 +18,7 @@ import java.util.Locale;
 import org.apache.jena.rdf.model.Resource;
 
 import io.openmanufacturing.sds.constraint.LocaleConstraint;
-import io.openmanufacturing.sds.metamodel.impl.DefaultLocaleConstraint;
+import io.openmanufacturing.sds.constraint.impl.DefaultLocaleConstraint;
 import io.openmanufacturing.sds.metamodel.loader.Instantiator;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.loader.ModelElementFactory;
