+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/metamodel/loader/instantiator/MeasurementInstantiator.java
@@ -21,7 +21,7 @@ import org.apache.jena.rdf.model.Statement;
 import io.openmanufacturing.sds.characteristic.Measurement;
 import io.openmanufacturing.sds.metamodel.Type;
 import io.openmanufacturing.sds.metamodel.Unit;
-import io.openmanufacturing.sds.metamodel.impl.DefaultMeasurement;
+import io.openmanufacturing.sds.characteristic.impl.DefaultMeasurement;
 import io.openmanufacturing.sds.metamodel.loader.Instantiator;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.loader.ModelElementFactory;
