+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/metamodel/loader/instantiator/RegularExpressionConstraintInstantiator.java
@@ -16,7 +16,7 @@ package io.openmanufacturing.sds.metamodel.loader.instantiator;
 import org.apache.jena.rdf.model.Resource;
 
 import io.openmanufacturing.sds.constraint.RegularExpressionConstraint;
-import io.openmanufacturing.sds.metamodel.impl.DefaultRegularExpressionConstraint;
+import io.openmanufacturing.sds.constraint.impl.DefaultRegularExpressionConstraint;
 import io.openmanufacturing.sds.metamodel.loader.Instantiator;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.loader.ModelElementFactory;
