+++ core/sds-aspect-meta-model-java/src/main/java/io/openmanufacturing/sds/metamodel/loader/instantiator/SingleEntityInstantiator.java
@@ -17,7 +17,7 @@ import org.apache.jena.rdf.model.Resource;
 
 import io.openmanufacturing.sds.characteristic.SingleEntity;
 import io.openmanufacturing.sds.metamodel.Type;
-import io.openmanufacturing.sds.metamodel.impl.DefaultSingleEntity;
+import io.openmanufacturing.sds.characteristic.impl.DefaultSingleEntity;
 import io.openmanufacturing.sds.metamodel.loader.Instantiator;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.metamodel.loader.ModelElementFactory;
