+++ core/sds-aspect-meta-model-java/src/test/java/io/openmanufacturing/sds/metamodel/loader/EnumerationInstantiatorTest.java
@@ -38,8 +38,8 @@ import io.openmanufacturing.sds.metamodel.ScalarValue;
 import io.openmanufacturing.sds.metamodel.Type;
 import io.openmanufacturing.sds.metamodel.Value;
 import io.openmanufacturing.sds.metamodel.datatypes.LangString;
-import io.openmanufacturing.sds.metamodel.impl.DefaultSet;
-import io.openmanufacturing.sds.metamodel.impl.DefaultSingleEntity;
+import io.openmanufacturing.sds.characteristic.impl.DefaultSet;
+import io.openmanufacturing.sds.characteristic.impl.DefaultSingleEntity;
 import io.openmanufacturing.sds.test.TestAspect;
 import io.openmanufacturing.sds.test.TestModel;
