+++ core/sds-aspect-model-java-generator/src/main/java/io/openmanufacturing/sds/aspectmodel/java/metamodel/StaticMetaModelJavaArtifactGenerator.java
@@ -77,28 +77,28 @@ import io.openmanufacturing.sds.metamodel.datatypes.LangString;
 import io.openmanufacturing.sds.metamodel.impl.BoundDefinition;
 import io.openmanufacturing.sds.metamodel.impl.DefaultAbstractEntity;
 import io.openmanufacturing.sds.metamodel.impl.DefaultCharacteristic;
-import io.openmanufacturing.sds.metamodel.impl.DefaultCode;
-import io.openmanufacturing.sds.metamodel.impl.DefaultCollection;
+import io.openmanufacturing.sds.characteristic.impl.DefaultCode;
+import io.openmanufacturing.sds.characteristic.impl.DefaultCollection;
 import io.openmanufacturing.sds.metamodel.impl.DefaultComplexType;
-import io.openmanufacturing.sds.metamodel.impl.DefaultDuration;
-import io.openmanufacturing.sds.metamodel.impl.DefaultEncodingConstraint;
+import io.openmanufacturing.sds.characteristic.impl.DefaultDuration;
+import io.openmanufacturing.sds.constraint.impl.DefaultEncodingConstraint;
 import io.openmanufacturing.sds.metamodel.impl.DefaultEntity;
-import io.openmanufacturing.sds.metamodel.impl.DefaultEnumeration;
-import io.openmanufacturing.sds.metamodel.impl.DefaultFixedPointConstraint;
-import io.openmanufacturing.sds.metamodel.impl.DefaultLanguageConstraint;
-import io.openmanufacturing.sds.metamodel.impl.DefaultLengthConstraint;
-import io.openmanufacturing.sds.metamodel.impl.DefaultList;
-import io.openmanufacturing.sds.metamodel.impl.DefaultMeasurement;
-import io.openmanufacturing.sds.metamodel.impl.DefaultQuantifiable;
-import io.openmanufacturing.sds.metamodel.impl.DefaultRangeConstraint;
-import io.openmanufacturing.sds.metamodel.impl.DefaultRegularExpressionConstraint;
+import io.openmanufacturing.sds.characteristic.impl.DefaultEnumeration;
+import io.openmanufacturing.sds.constraint.impl.DefaultFixedPointConstraint;
+import io.openmanufacturing.sds.constraint.impl.DefaultLanguageConstraint;
+import io.openmanufacturing.sds.constraint.impl.DefaultLengthConstraint;
+import io.openmanufacturing.sds.characteristic.impl.DefaultList;
+import io.openmanufacturing.sds.characteristic.impl.DefaultMeasurement;
+import io.openmanufacturing.sds.characteristic.impl.DefaultQuantifiable;
+import io.openmanufacturing.sds.constraint.impl.DefaultRangeConstraint;
+import io.openmanufacturing.sds.constraint.impl.DefaultRegularExpressionConstraint;
 import io.openmanufacturing.sds.metamodel.impl.DefaultScalar;
-import io.openmanufacturing.sds.metamodel.impl.DefaultSet;
-import io.openmanufacturing.sds.metamodel.impl.DefaultSingleEntity;
-import io.openmanufacturing.sds.metamodel.impl.DefaultSortedSet;
-import io.openmanufacturing.sds.metamodel.impl.DefaultState;
-import io.openmanufacturing.sds.metamodel.impl.DefaultStructuredValue;
-import io.openmanufacturing.sds.metamodel.impl.DefaultTrait;
+import io.openmanufacturing.sds.characteristic.impl.DefaultSet;
+import io.openmanufacturing.sds.characteristic.impl.DefaultSingleEntity;
+import io.openmanufacturing.sds.characteristic.impl.DefaultSortedSet;
+import io.openmanufacturing.sds.characteristic.impl.DefaultState;
+import io.openmanufacturing.sds.characteristic.impl.DefaultStructuredValue;
+import io.openmanufacturing.sds.characteristic.impl.DefaultTrait;
 import io.openmanufacturing.sds.metamodel.loader.MetaModelBaseAttributes;
 import io.openmanufacturing.sds.staticmetamodel.PropertyContainer;
 import io.openmanufacturing.sds.staticmetamodel.StaticContainerProperty;
