+++ core/sds-aspect-model-java-generator/src/main/java/velocity_implicit.vm
@@ -22,29 +22,29 @@
 #* @vtlvariable name="dataType" type="java.lang.String" *#
 #* @vtlvariable name="deconstructor" type="io.openmanufacturing.sds.aspectmodel.java.StructuredValuePropertiesDeconstructor" *#
 #* @vtlvariable name="DefaultCharacteristic" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultCharacteristic>" *#
-#* @vtlvariable name="DefaultCode" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultCode>" *#
-#* @vtlvariable name="DefaultCollection" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultCollection>" *#
-#* @vtlvariable name="DefaultDuration" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultDuration>" *#
-#* @vtlvariable name="DefaultEncodingConstraint" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultEncodingConstraint>" *#
+#* @vtlvariable name="DefaultCode" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultCode>" *#
+#* @vtlvariable name="DefaultCollection" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultCollection>" *#
+#* @vtlvariable name="DefaultDuration" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultDuration>" *#
+#* @vtlvariable name="DefaultEncodingConstraint" type="java.lang.Class<io.openmanufacturing.sds.constraint.impl.DefaultEncodingConstraint>" *#
 #* @vtlvariable name="DefaultEntity" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultEntity>" *#
-#* @vtlvariable name="DefaultEnumeration" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultEnumeration>" *#
-#* @vtlvariable name="DefaultFixedPointConstraint" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultFixedPointConstraint>" *#
-#* @vtlvariable name="DefaultLanguageConstraint" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultLanguageConstraint>" *#
-#* @vtlvariable name="DefaultLocaleConstraint" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultLocaleConstraint>" *#
-#* @vtlvariable name="DefaultLengthConstraint" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultLengthConstraint>" *#
-#* @vtlvariable name="DefaultList" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultList>" *#
-#* @vtlvariable name="DefaultMeasurement" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultMeasurement>" *#
-#* @vtlvariable name="DefaultQuantifiable" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultQuantifiable>" *#
-#* @vtlvariable name="DefaultRangeConstraint" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultRangeConstraint>" *#
-#* @vtlvariable name="DefaultRegularExpressionConstraint" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultRegularExpressionConstraint>" *#
+#* @vtlvariable name="DefaultEnumeration" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultEnumeration>" *#
+#* @vtlvariable name="DefaultFixedPointConstraint" type="java.lang.Class<io.openmanufacturing.sds.constraint.impl.DefaultFixedPointConstraint>" *#
+#* @vtlvariable name="DefaultLanguageConstraint" type="java.lang.Class<io.openmanufacturing.sds.constraint.impl.DefaultLanguageConstraint>" *#
+#* @vtlvariable name="DefaultLocaleConstraint" type="java.lang.Class<io.openmanufacturing.sds.constraint.impl.DefaultLocaleConstraint>" *#
+#* @vtlvariable name="DefaultLengthConstraint" type="java.lang.Class<io.openmanufacturing.sds.constraint.impl.DefaultLengthConstraint>" *#
+#* @vtlvariable name="DefaultList" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultList>" *#
+#* @vtlvariable name="DefaultMeasurement" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultMeasurement>" *#
+#* @vtlvariable name="DefaultQuantifiable" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultQuantifiable>" *#
+#* @vtlvariable name="DefaultRangeConstraint" type="java.lang.Class<io.openmanufacturing.sds.constraint.impl.DefaultRangeConstraint>" *#
+#* @vtlvariable name="DefaultRegularExpressionConstraint" type="java.lang.Class<io.openmanufacturing.sds.constraint.impl.DefaultRegularExpressionConstraint>" *#
 #* @vtlvariable name="DefaultScalar" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultScalar>" *#
 #* @vtlvariable name="DefaultScalarValue" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultScalarValue>" *#
-#* @vtlvariable name="DefaultSet" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultSet>" *#
-#* @vtlvariable name="DefaultSingleEntity" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultSingleEntity>" *#
-#* @vtlvariable name="DefaultSortedSet" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultSortedSet>" *#
-#* @vtlvariable name="DefaultState" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultState>" *#
-#* @vtlvariable name="DefaultStructuredValue" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultStructuredValue>" *#
-#* @vtlvariable name="DefaultTrait" type="java.lang.Class<io.openmanufacturing.sds.metamodel.impl.DefaultTrait>" *#
+#* @vtlvariable name="DefaultSet" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultSet>" *#
+#* @vtlvariable name="DefaultSingleEntity" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultSingleEntity>" *#
+#* @vtlvariable name="DefaultSortedSet" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultSortedSet>" *#
+#* @vtlvariable name="DefaultState" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultState>" *#
+#* @vtlvariable name="DefaultStructuredValue" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultStructuredValue>" *#
+#* @vtlvariable name="DefaultTrait" type="java.lang.Class<io.openmanufacturing.sds.characteristic.impl.DefaultTrait>" *#
 #* @vtlvariable name="Duration" type="java.lang.Class<io.openmanufacturing.sds.characteristic.Duration>" *#
 #* @vtlvariable name="Either" type="java.lang.Class<io.openmanufacturing.sds.aspectmodel.java.types.Either>" *#
 #* @vtlvariable name="element" type="io.openmanufacturing.sds.metamodel.StructureElement" *#
