+++ core/sds-aspect-model-java-generator/src/test/java/io/openmanufacturing/sds/aspectmodel/java/StaticMetaModelJavaGeneratorTest.java
@@ -40,8 +40,8 @@ import com.google.common.reflect.TypeToken;
 import io.openmanufacturing.sds.aspectmetamodel.KnownVersion;
 import io.openmanufacturing.sds.metamodel.datatypes.Curie;
 import io.openmanufacturing.sds.metamodel.impl.DefaultCharacteristic;
-import io.openmanufacturing.sds.metamodel.impl.DefaultList;
-import io.openmanufacturing.sds.metamodel.impl.DefaultMeasurement;
+import io.openmanufacturing.sds.characteristic.impl.DefaultList;
+import io.openmanufacturing.sds.characteristic.impl.DefaultMeasurement;
 import io.openmanufacturing.sds.staticmetamodel.StaticContainerProperty;
 import io.openmanufacturing.sds.staticmetamodel.StaticProperty;
 import io.openmanufacturing.sds.staticmetamodel.StaticUnitProperty;
